Reorg functions, extract pure functions to utils and update JSDoc types
